ecommerce_project/requirements.txt
asgiref==3.11.1
click==8.3.2
colorama==0.4.6
dj-database-url==3.1.2
Django==5.2.13
django-cors-headers==4.9.0
djangorestframework==3.17.1
dlib==20.0.1
face-recognition==1.3.0
face-recognition-models==0.3.0
gunicorn==25.3.0
numpy==2.2.6
opencv-python==4.13.0.92
packaging==26.1
pillow==12.2.0
psycopg2-binary==2.9.11
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2026.1
whitenoise==6.12.0
